@@ -1,5 +1,2 @@
-from typing import Tuple
-
-
-def get_zone_filenames(zone: int) -> Tuple[str, str]:
+def get_zone_filenames(zone: int) -> tuple[str, str]:
     return f'{zone:03d}1.VGA', f'{zone:03d}2.VGA'
